refactor(tests_: separate tests by responsibility
