+++ __tests__/mock/response/remove-attributes.js
