+++ __tests__/mock/response/remove-data.js
