+++ __tests__/response/remove-attributes.spec.js
@@ -1,42 +1,42 @@
 'use strict';
 
-const { modifyResponseBodyData } = require('../server/services/transform-service/response');
-const { initial, removeDataKey } = require('./mock');
+const { modifyResponseBodyData } = require('../../server/services/transform-service/response');
+const { initial, removeAttributesKey } = require('../mock/response');
 
-const transformOptions = { removeDataKey: true };
+const transformOptions = { removeAttributesKey: true };
 
-describe('removeDataKey', () => {
+describe('removeAttributesKey', () => {
 	// single relation
 	test('object with data property of type object', () => {
 		const result = modifyResponseBodyData(transformOptions, initial.dataObject);
 		expect(result).toBeDefined();
-		expect(result).toEqual(removeDataKey.dataObject);
+		expect(result).toEqual(removeAttributesKey.dataObject);
 	});
 
 	test('object with data property of type null', () => {
 		const result = modifyResponseBodyData(transformOptions, initial.dataWithNull);
 		expect(result).toBeDefined();
-		expect(result).toEqual(removeDataKey.dataWithNull);
+		expect(result).toEqual(removeAttributesKey.dataWithNull);
 	});
 
 	// multiple relation
 	test('object with data property of type array', () => {
 		const result = modifyResponseBodyData(transformOptions, initial.dataArray);
 		expect(result).toBeDefined();
-		expect(result).toEqual(removeDataKey.dataArray);
+		expect(result).toEqual(removeAttributesKey.dataArray);
 	});
 
 	test('object with data property of type array with no length', () => {
 		const result = modifyResponseBodyData(transformOptions, initial.dataWithEmptyArray);
 		expect(result).toBeDefined();
-		expect(result).toEqual(removeDataKey.dataWithEmptyArray);
+		expect(result).toEqual(removeAttributesKey.dataWithEmptyArray);
 	});
 
 	// single component
 	test('object with id property', () => {
 		const result = modifyResponseBodyData(transformOptions, initial.id);
 		expect(result).toBeDefined();
-		expect(result).toEqual(removeDataKey.id);
+		expect(result).toEqual(removeAttributesKey.id);
 	});
 
 	// multiple components
@@ -44,20 +44,20 @@ describe('removeDataKey', () => {
 	test('array of objects containing an id property', () => {
 		const result = modifyResponseBodyData(transformOptions, initial.arrayWithIds);
 		expect(result).toBeDefined();
-		expect(result).toEqual(removeDataKey.arrayWithIds);
+		expect(result).toEqual(removeAttributesKey.arrayWithIds);
 	});
 
 	// skip single media
 	test('object with provider property', () => {
 		const result = modifyResponseBodyData(transformOptions, initial.provider);
 		expect(result).toBeDefined();
-		expect(result).toEqual(removeDataKey.provider);
+		expect(result).toEqual(removeAttributesKey.provider);
 	});
 
 	// skip multi media
 	test('array of objects containing a provider property', () => {
 		const result = modifyResponseBodyData(transformOptions, initial.arrayWithProviders);
 		expect(result).toBeDefined();
-		expect(result).toEqual(removeDataKey.arrayWithProviders);
+		expect(result).toEqual(removeAttributesKey.arrayWithProviders);
 	});
 });
